QueryDsl

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -141,7 +141,7 @@ class BookServiceTest @Autowired constructor(
 //        assertThat(scienceDto.count).isEqualTo(1)
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
 }
